requirements.txt
altgraph==0.17.2
asgiref==3.4.1
auto-py-to-exe==2.11.0
backports.entry-points-selectable==1.1.1
boto3==1.6.23
botocore==1.9.23
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
distlib==0.3.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.1
django-appconf==1.0.2
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-html==0.1.0
django-js-asset==2.0.0
django-multiselectfield==0.1.8
django-phonenumber-field==6.0.0
django-staticfiles==1.2.1
django-storages==1.6.6
django-widget-tweaks==1.4.9
django-wkhtmltopdf==3.2.0
djangorestframework==3.13.1
docutils==0.14
Eel==0.12.4
filelock==3.3.2
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==19.7.1
jmespath==0.9.3
Markdown==3.3.6
markdown2==2.4.2
pdfkit==1.0.0
pefile==2021.9.3
phonenumbers==8.12.40
Pillow==8.3.2
pipenv==2021.11.23
platformdirs==2.4.0
Pmw==2.0.1
psycopg2==2.9.3
pycparser==2.21
pyglet==1.5.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.3
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.6.1
python-decouple==3.6
pytz==2018.3
pywin32-ctypes==0.2.0
reportlab==3.6.2
s3transfer==0.1.13
six==1.11.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
unrar==0.4
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
whichcraft==0.6.1
whitenoise==5.2.0
wkhtmltopdf==0.2
zope.event==4.5.0
zope.interface==5.4.0
